+++ AuleWeb/src/main/java/data/domainImpl/AmministratoreImpl.java
@@ -10,7 +10,33 @@ import data.domain.Amministratore;
  *
  * @author alessandro
  */
-public class AmministratoreImpl extends UtenteImpl implements Amministratore{
+public class AmministratoreImpl implements Amministratore{
+
+    
+    private String username;
+    
+    private String password;
+    
+    
+    @Override
+    public String getUseraname() {
+        return this.username;
+    }
+
+    @Override
+    public String getPassword() {
+        return this.password;
+    }
+
+    @Override
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    @Override
+    public void setPassword(String password) {
+        this.password = password;
+    }
     
     
 }
